pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "better-finder-mcp"
version = "0.1.0"
description = "Enhanced macOS finder using MCP with intelligent file indexing and retrieval"
authors = [{name = "zeroualihamza0206@gmail.com"}]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "mcp>=1.0.0",
    "rich>=13.0.0",
    "faiss-cpu>=1.7.0",
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "sentence-transformers>=2.7.0",
    "PyPDF2>=3.0.0",
    "openpyxl>=3.1.0",
    "python-docx>=1.1.0",
    "aiofiles>=24.0.0",
    "watchdog>=4.0.0",
    "pydantic>=2.0.0",
    "typer>=0.12.0",
    "asyncio-mqtt>=0.16.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "ruff>=0.1.0"
]

[project.scripts]
better-finder = "enhanced_finder.main:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["enhanced_finder*"]

[tool.black]
line-length = 88
target-version = ['py39']

[tool.ruff]
line-length = 88
target-version = "py39"